refactor: 이동 조건 관련 패키지 utils로 분리
