+++ src/main/java/racingcar/utils/MovableCondition.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.utils;
 
 public interface MovableCondition {
     boolean isMovable();
