+++ src/main/java/racingcar/utils/RandomValueGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.utils;
 
 @FunctionalInterface
 public interface RandomValueGenerator {
